+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- Mechanism to resolve absolute paths in include directives
+
 ## 1.1.0 - 2017-02-24
 ### Added
 - Mechanism to follow include directives to find all declared taglibs